05_mail/requirements.txt
# 見積依頼スキル 依存ライブラリ

# Outlook連携
pywin32>=306

# メールアドレス検証
email-validator>=2.0.0

# 文字コード自動判定
chardet>=5.0.0

# 暗号化
cryptography>=41.0.0

# Windows資格情報マネージャー
keyring>=24.0.0

# Wordテンプレート処理
python-docx>=1.0.0

# URL有効性チェック
requests>=2.31.0
